FogBugz #103239: Intermittent loss of ethernet transmission.
This bug does not really have anything to do with ethernet.
The problem was a memory coherency problem between the cpu
and a dma engine, in this case the ethernet mac.  In short,
while dma writes by the ethernet were not getting cached in L2,
they were getting cached in L1.  The good news is a single bit
change to the programming of the PL310 fixes the problem.

Signed-off-by: Matthew Gerlach <mgerlach@altera.com>