FogBugz #168684: Disable NAND driver in devicetree
On our CycloneV, sometimes a NAND chip is detected even though
there isn't one on the board.  The Denali NAND driver already
has code to prevent erroneous detections, but apparently it's
not working as well as it could.  Our solution is to disable the NAND
driver by default for socfpga.

Signed-off-by: Graham Moore <grmoore@altera.com>