FogBugz #171833: Strip VLAN tags in EMAC Receive path so GRO works as expected
The EMAC driver was advertising the NETIF_F_HW_VLAN_TAG_RX feature flag,
which indicates that the hardware supports VLAN tag stripping in receive
frames, but the hardware does not support VLAN tag stripping, and the
driver was not stripping the VLANs either. This causes a slow processing
path to be taken in the GRO portion of the stack's receive processing,
drastically impacting receive performance. The fix is for the driver to
detect and strip the VLAN tag correctly, and correctly setting any
detected and received VLAN tag into the skb.

Signed-off-by: Vince Bridgers <vbridger@altera.com>

Conflicts:

	drivers/net/ethernet/stmicro/stmmac/stmmac_main.c