FogBugz #237341: fix fetching cpu1start_addr for SMP
When CPU1 is brought out of reset, it's MMU is not turned on yet, so it
will only be able to use physical addresses. For systems that have their
MMU page configured for 0xC0000000, 0x80000000, or 0x40000000
"BIC 0x40000000" will work just fine, as it was just converting the
virtual address of &cpu1start_addr into a physical address, ie. 0xC0000000
became 0x80000000. So for systems where the SDRAM controller was able to do
a wrap-around access, this was working fine, as it was just dropping the
MSB, but for systems where out of bounds memory access is not allowed, this
would not allow CPU1 to correctly fetch &cpu1start_addr.

This patch fixes the secondary_trampoline code to correctly fetch the
physical address of cpu1start_addr directly. The patch will grab the
correct PAGE_OFFSET and convert &cpu1start_addr to the physical address
correctly.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>