FogBugz #314791: QSPI: Set jedec_id in flash data structure
This patch initializes the jedec_id in the flash data structure
so that the write_ear() function will send the correct
bank-select command to the device.  This patch is only for kernel
version 3.10-ltsi.

Mentor Graphics reported a bug whereby the extended address register
command is not setup properly, and therefore the write_ear()
function sends an ill-formed command to the device, causing the
Arria10 Virtual Platform QSPI interface to fail.

This has not been a problem on real hardware, probably because the
chip-select is deasserted in the middle of the bad command, which
essentially aborts the command.  The Arria10 Virtual Platform
doesn't have a chip-select, therefore the bad command is not
aborted, and the VP cannot recover.

Furthermore, the driver is not actually using a bank-select
method of addressing, it is using 4-byte address mode, and
the extended address register is ignored by the device.  The
write_ear() function has been removed from later kernel versions.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>