FogBugz #317354: EDAC driver should not look for ECC_CORR_EN
The decision was made by the bootloader to not support ECC_CORR_EN
functionality. Disabling ECC_CORR_EN, and only leaving ECC_EN means
that when error happens, it is the user's responsibility to perform
a full SDRAM scrub.

So for now, we will not be supporting ECC_CORR_EN in the EDAC driver.

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>