FogBugz #350137: Fix A10SyCon Button Export Crash
Exporting the A10Sycon buttons would cause a kernel crash. The root
cause was that the interrupt was not mapped but was being accessed.
In part this was because the GPIO bank that handled the Button was
disabled but this should be handled gracefully.

Check for a valid IRQ each time and exit if not valid.

Signed-off-by: Thor Thayer <tthayer@opensource.altera.com>