FogBugz #352163-2: fix error checking in interrupt latency driver
The ZERO-day testing revealed an error in how a variable is declared.
The variable was declared as an unsigned int, but the error check if
a value < 0. This will never happen if the variable is an unsigned int.

An addendum to commit 01740330db6f ("FogBugz #352163: fix error checking
in interrupt latency driver").

Reported-by: kbuild test robot <fengguang.wu@intel.com>
Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>