FogBugz #355420-2: dts: socfpga: Fix spidev WARN_ON() during boot.
A WARN_ON() was added to the spidev framework and the spidef compatible
string was removed. The preference is to use an actual part. In our
case, the SPI port is unused but still needs to be tested with each
release. The fix is to spoof spidev using the existing ROHM DAC
compatible string.

Signed-off-by: Thor Thayer <tthayer@opensource.altera.com>