FogBugz #383074-1: fpga-mgr: add fpga image information struct
A FPGA design contained in a FPGA image file will likely have
particulars that affect how the image is programmed to the FPGA.
Currently the only such particular is a single flag bit
indicating whether the image is for full or partial reconfig-
uration.

This patch adds a new 'struct  fpga_image_info', moves the
current 'u32 flags' to it.

The FPGA Manager API functions are changed, as the 'u32 flag'
parameter is replaced with a pointer to struct fpga_image_info.

Subsequent patches will add more to the struct.  The things
that will be added for current use are timeouts for knowing how
long to reasonably wait when shutting down/re-enabling traffic
to a FPGA region that is reprogrammed.  That can vary widely
with the implementation of the design that is being programmed.
Other future additions will be to handle passing addresses for
reprogramming relocatable FPGA images.  As use of FPGA in the
kernel continues to develop, other specifics are likely to be
added.

Signed-off-by: Alan Tull <atull@opensource.altera.com>