FogBugz #383074-2: fpga-bridge: add fpga image information struct
Pass the FPGA Manager framework's 'struct  fpga_image_info'
around for use of the FPGA Bridges during FPGA reprogramming.

Add two timeouts to fpga_image_info for knowing how long to
wait when shutting down/re-enabling traffic to a FPGA region
that is reprogrammed.  That can vary widely with the
implementation of the design that is being programmed.

Signed-off-by: Alan Tull <atull@opensource.altera.com>